Update tokeniser & allow library to use big buffers
In fact the big buffers is quite a significant thing.  On 64 bit
systems, the mmap() size is 64 bit, so all Tracter buffers should have
64 bit sizes.  Hence, there is now a SizeType to define this.

+++ src/MMap.cpp
@@ -60,7 +60,7 @@ void* Tracter::MMap::Map(const char* iFileName)
     mSize = buf.RegionSize;
 
     if (sVerbose > 1)
-        printf("MMap: %s size %d\n", iFileName, (int)mSize);
+        printf("MMap: %s size %ld\n", iFileName, (long int)mSize);
 
     return mMap;
 }
@@ -135,7 +135,7 @@ void* Tracter::MMap::Map(const char* iFileName)
     }
 
     if (sVerbose > 1)
-        printf("MMap: %s size %d\n", iFileName, (int)mSize);
+        printf("MMap: %s size %ld\n", iFileName, (long int)mSize);
 
     return mMap;
 }
